separate bucket manager from main
